+++ src/controller_device_handle.h
@@ -116,6 +116,7 @@ char  *device_handle_domain_get(device_handle dh);
 int    device_handle_domain_set(device_handle dh, char *domain);
 cbuf  *device_handle_outmsg_get(device_handle dh, int nr);
 int    device_handle_outmsg_set(device_handle dh, int nr, cbuf *cb);
+int    device_handle_stats(clixon_handle  h, uint64_t *nrp, size_t *szp);
 
 #ifdef __cplusplus
 }